MA-24240-1 Add output buffer with RGBA8888 support for virtual display
When the formatHint of virtual display is RGBA8888, no convert buffer allocated.
Need to process the case that the output buffer format does not match with
formatHint of virtual display.

Reviewed-on: http://androidsource.nxp.com/project/30655

+++ display/hwcomposer3/Common.h
@@ -199,7 +199,8 @@ inline bool rectIntersect(const common::Rect& with, common::Rect& result) {
 
 inline uint64_t calculateRect(common::Rect& rect) {
     if ((rect.left < rect.right) && (rect.top < rect.bottom))
-        return (rect.right - rect.left) * (rect.bottom - rect.top);
+        return ((uint64_t)rect.right - (uint64_t)rect.left) *
+                ((uint64_t)rect.bottom - (uint64_t)rect.top);
     else
         return 0;
 }
